move metrics usage out of program-runtime
