+++ program-runtime-metrics/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "solana-program-runtime-metrics"
+description = "Metrics utilities for solana-program-runtime"
+documentation = "https://docs.rs/solana-program-runtime-metrics"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+log = { workspace = true }
+solana-measure = { workspace = true }
+solana-metrics = { workspace = true }
+solana-program-runtime = { workspace = true }
+solana-pubkey = { workspace = true }
+solana_rbpf = { workspace = true }
+solana-sdk = { workspace = true }
+solana-timings = { workspace = true }
+solana-type-overrides = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
