+++ program-runtime/src/lib.rs
@@ -2,9 +2,6 @@
 #![deny(clippy::arithmetic_side_effects)]
 #![deny(clippy::indexing_slicing)]
 
-#[macro_use]
-extern crate solana_metrics;
-
 pub use solana_rbpf;
 pub mod invoke_context;
 pub mod loaded_programs;
