+++ programs/loader-v4/Cargo.toml
@@ -15,6 +15,7 @@ solana-compute-budget = { workspace = true }
 solana-log-collector = { workspace = true }
 solana-measure = { workspace = true }
 solana-program-runtime = { workspace = true }
+solana-program-runtime-metrics = { workspace = true }
 solana-sdk = { workspace = true }
 solana-type-overrides = { workspace = true }
 solana_rbpf = { workspace = true }
